+++ Woof/Woof/Helper/AppStyle/AppStyle+UIElementConstant.swift
@@ -14,5 +14,8 @@ extension AppStyle {
 
         /// The default side size of an avatar image.
         static let avatarSideSize = 80.0
+
+        /// The default spacing that can be used between elements in View.
+        static let spacingBetweenElements = 8.0
     }
 }