+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added 
 
 - Add `OwnerCardView` as part of `OwnerInformationView`, `OwnerInformationView` to the Profile's tab in `MainTabView` - [#47](https://github.com/ios-course/ironfoudation-team-project/pull/47)
+- Add `OwnerCardView`- [81](https://github.com/ios-course/ironfoudation-team-project/pull/81)
 
 ## [0.1.0] - 31.05.2023
 
@@ -30,4 +31,3 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Add missed information to `CHANGELOG.md`  - [#56](https://github.com/ios-course/ironfoudation-team-project/pull/56)
 - Add the `KeyValueStorage` class to manage persistent data storage - [58](https://github.com/ios-course/ironfoudation-team-project/pull/58)
 - Add `EditOwnerInformationView` - [77](https://github.com/ios-course/ironfoudation-team-project/pull/77)
-- Add `OwnerCardView`- [81](https://github.com/ios-course/ironfoudation-team-project/pull/81)