+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added 
 
 - Add `OwnerProfileViewModel` for `OwnerProfileView` - [#88](https://github.com/ios-course/ironfoudation-team-project/pull/88)
+- Add `OwnerCardView`- [81](https://github.com/ios-course/ironfoudation-team-project/pull/81)
 
 ## [0.1.0] - 31.05.2023